requirements.txt
aiohappyeyeballs==2.4.6
aiohttp==3.11.13
aiolimiter==1.2.1
aiosignal==1.3.2
annotated-types==0.7.0
anthropic==0.48.0
anyio==4.8.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
attrs==25.1.0
cachetools==5.5.2
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
dashscope==1.22.1
distro==1.9.0
eval_type_backport==0.2.2
frozenlist==1.5.0
google-ai-generativelanguage==0.6.15
google-api-core==2.24.1
google-api-python-client==2.162.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-genai==1.3.0
google-generativeai==0.8.4
googleapis-common-protos==1.68.0
grpcio==1.71.0rc2
grpcio-status==1.71.0rc2
h11==0.14.0
httpcore==1.0.7
httplib2==0.22.0
httpx==0.28.1
idna==3.10
jiter==0.8.2
jsonpath-python==1.0.6
loguru==0.7.3
minio==7.2.15
mistralai==1.5.0
multidict==6.1.0
mypy-extensions==1.0.0
numpy==2.2.3
openai==1.64.0
packaging==24.2
pillow==11.1.0
propcache==0.3.0
proto-plus==1.26.0
protobuf==5.29.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pycryptodome==3.21.0
pydantic==2.10.6
pydantic_core==2.27.2
pyparsing==3.2.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
PyYAML==6.0.2
replicate==1.0.4
requests==2.32.3
rsa==4.9
scipy==1.15.2
six==1.17.0
sniffio==1.3.1
tqdm==4.67.1
typing-inspect==0.9.0
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.3.0
websocket-client==1.8.0
websockets==14.2
yarl==1.18.3
